Underlying indicators for Scottish Index of Multiple Deprivation 2012

A data set containing proportions and extents for Council Areas (Local Authorities) in Scotland. Note that IMD scores for Scotland are not available.

Usage

imd2012_scotland_dz01_indicators

Format

A data frame of class "tbl" with 6,505 rows and 47 variables:

dz01_code

dz01_code

total_population_sape_2010

Total Population (SAPE 2010)

income_domain_2012_rank

Income domain 2012 rank

income_domain_2012_rate_percent

Income domain 2012 rate (percent)

number_of_income_deprived_people_2012

Number of Income Deprived People 2012

best_fit_working_age_population_men_16_64_women_16_60_sape_2010

Best-fit Working Age Population (men 16-64, women 16-60 SAPE 2010)

employment_domain_2012_rank

Employment domain 2012 rank

employment_domain_2012_rate_percent

Employment domain 2012 rate (percent)

number_of_employment_deprived_people_2012

Number of Employment Deprived People 2012

health_domain_2012_rank

Health domain 2012 rank

health_domain_2012_score1

Health domain 2012 score1

standardised_mortality_ratio_isd_2007_2010

Standardised mortality ratio (ISD, 2007-2010)

comparative_illness_factor_standardised_ratio_dwp_2011_2

Comparative illness factor: standardised ratio (DWP, 2011)2

hospital_stays_related_to_alcohol_misuse_standardised_ratio_isd_2007_2010

Hospital stays related to alcohol misuse: standardised ratio(ISD, 2007-2010)

hospital_stays_related_to_drug_misuse_standardised_ratio_isd_2007_2010

Hospital stays related to drug misuse: standardised ratio (ISD, 2007-2010)

emergency_stays_in_hospital_standardised_ratio_isd_2007_2010

Emergency stays in hospital: standardised ratio (ISD, 2007-2010)

estimated_proportion_of_population_being_prescribed_drugs_for_anxiety_depression_or_psychosis_isd_2010

Estimated proportion of population being prescribed drugs for anxiety, depression or psychosis (ISD, 2010)

proportion_of_live_singleton_births_of_low_birth_weight_isd_2006_2009

Proportion of live singleton births of low birth weight(ISD, 2006-2009)

education_skills_and_training_domain_2012_rank1

Education, Skills and Training domain 2012 rank1

education_skills_and_training_domain_2012_score1

Education, Skills and Training domain 2012 score1

working_age_people_with_no_qualifications_2001_2

Working age people with no qualifications (2001)2

people_aged_16_19_not_in_full_time_education_employment_or_training_rate_school_leavers_2009_10_2010_11_dwp_2010_and_2011

People aged 16-19 not in full time education, employment or training rate (School Leavers 2009/10-2010-11, DWP 2010 and 2011)

proportion_of_17_21_year_olds_entering_higher_education_hesa_2008_09_2010_11_3

Proportion of 17- 21 year olds entering higher education (HESA 2008/09-2010/11)3

pupil_performance_on_sqa_at_stage_4_sqa_2008_09_2010_11_5

Pupil Performance on SQA at Stage 4 (SQA, 2008/09-2010/11)5

school_pupil_absences_scottish_government_2009_10_2010_11_4_5

School Pupil Absences (Scottish Government, 2009/10-2010/11)4,5

housing_domain_rank_2004_2006_2009_2012

Housing domain rank 2004, 2006, 2009 & 2012

housing_domain_score_2004_2006_2009_2012

Housing domain score 2004, 2006, 2009 & 2012

household_population_census_2001

Household Population (Census, 2001)

percentage_of_household_population_living_in_households_without_central_heating_census_2001

Percentage of household population living in households without central heating (Census, 2001)

percentage_of_household_population_living_in_households_that_are_overcrowded_census_2001

Percentage of household population living in households that are overcrowded (Census, 2001)

geographic_access_domain_2012_rank

Geographic Access domain 2012 rank

geographic_access_domain_2012_score

Geographic Access domain 2012 score

drive_times_sub_domain_2012_rank

Drive times sub-domain 2012 rank

public_transport_sub_domain_2012_rank

Public transport sub-domain 2012 rank

drive_time_to_gp_2012_mins

Drive time to GP 2012 (mins)

drive_time_to_petrol_station_2012_mins

Drive time to Petrol Station 2012 (mins)

drive_time_to_post_office_2012_mins

Drive time to Post Office 2012 (mins)

drive_time_to_primary_school_2012_mins

Drive time to Primary School 2012 (mins)

drive_time_to_secondary_school_2012_mins

Drive time to Secondary School 2012 (mins)

drive_time_to_retail_centre_2012_mins

Drive time to retail centre 2012 (mins)

public_transport_travel_time_to_gp_2012_mins

Public transport travel time to GP 2012 (mins)

public_transport_travel_time_to_post_office_2012_mins

Public transport travel time to Post Office 2012 (mins)

public_transport_travel_time_to_retail_centre_2012_mins

Public transport travel time to retail centre 2012 (mins)

police_force_area

Police Force Area

simd_crime_2012_rank

SIMD Crime 2012 rank

simd_crime_2012_count

SIMD Crime 2012 count

simd_crimes_per_10_000_total_population

SIMD Crimes per 10,000 total population

...

Source

https://www.gov.scot/