Underlying indicators for Scottish Index of Multiple Deprivation 2020

A data set containing proportions and extents for Council Areas (Local Authorities) in Scotland. Note that IMD scores for Scotland are not available.

Usage

imd2020_scotland_dz11_indicators

Format

A data frame of class "tbl" with 6,976 rows and 35 variables:

dz11_code

Data Zone code

Total_population

Total population

Working_age_population

Working age population

Income_rate

Percentage of people who are income deprived

Income_count

Number of people who are income deprived

Employment_rate

Percentage of people who are employment deprived

Employment_count

Number of people who are employment deprived

CIF

Comparative Illness Factor: standardised ratio

ALCOHOL

Hospital stays related to alcohol misuse: standardised ratio

DRUG

Hospital stays related to drug misuse: standardised ratio

SMR

Standardised mortality ratio

DEPRESS

Proportion of population being prescribed drugs for anxiety, depression or psychosis

LBWT

Proportion of live singleton births of low birth weight

EMERG

Emergency stays in hospital: standardised ratio

Attendance

School pupil attendance

Attainment

Attainment of school leavers

no_qualifications

Working age people with no qualifications: standardised ratio

not_participating

Proportion of people aged 16-19 not participating in education, employment or training

University

Proportion of 17-21 year olds entering university

drive_petrol

Average drive time to a petrol station in minutes

drive_GP

Average drive time to a GP surgery in minutes

drive_post

Average drive time to a post office in minutes

drive_primary

Average drive time to a primary school in minutes

drive_retail

Average drive time to a retail centre in minutes

drive_secondary

Average drive time to a secondary school in minutes

PT_GP

Public transport travel time to a GP surgery in minutes

PT_post

Public transport travel time to a post office in minutes

PT_retail

Public transport travel time to a retail centre in minutes

Broadband

Percentage of premises without access to superfast broadband (at least 30Mb/s download speed)

crime_count

Number of recorded crimes of violence, sexual offences, domestic housebreaking, vandalism, drugs offences, and common assault

crime_rate

Recorded crimes of violence, sexual offences, domestic housebreaking, vandalism, drugs offences, and common assault per 10,000 people

overcrowded_count

Number of people in households that are overcrowded

nocentralheat_count

Number of people in households without central heating

overcrowded_rate

Percentage of people in households that are overcrowded

nocentralheat_rate

Percentage of people in households without central heating

...

Source

https://www.gov.scot/