Index of Multiple Deprivation (2025) Indicators for LSOAs (2021) in England

A dataset containing Index of Multiple Deprivation indicators for Lower Layer Super Output Areas (LSOAs) in England.

Usage

imd2025_england_lsoa21_indicators

Format

A data frame of class "tbl":

lsoa21_code

LSOA code (2021)

income_domain_numerator

Income Domain numerator

income_deprivation_affecting_children_index_idaci_numerator

Income Deprivation Affecting Children Index (IDACI) numerator

income_deprivation_affecting_older_people_index_idaopi_numerator

Income Deprivation Affecting Older People Index (IDAOPI) numerator

employment_domain_numerator

Employment Domain numerator

adult_skills_and_english_language_proficiency_indicator

Adult skills and English language proficiency indicator

entry_to_higher_education_indicator

Entry to higher education indicator

acute_morbidity_indicator

Acute morbidity indicator

comparative_illness_and_disability_ratio_indicator

Comparative illness and disability ratio indicator

mental_health_indicator

Mental health indicator

years_of_potential_life_lost_indicator

Years of potential life lost indicator

violence_with_injury_rate_per_1_000_at_risk_population

Violence with injury, rate per 1,000 'at-risk population'

violence_without_injury_rate_per_1_000_at_risk_population

Violence without injury, rate per 1,000 'at-risk population'

stalking_and_harassment_rate_per_1_000_at_risk_population

Stalking and harassment, rate per 1,000 'at-risk population'

burglary_rate_per_1_000_at_risk_properties

Burglary, rate per 1,000 'at-risk properties'

theft_rate_per_1_000_at_risk_population

Theft, rate per 1,000 'at-risk population'

criminal_damage_rate_per_1_000_at_risk_population

Criminal damage, rate per 1,000 'at-risk population'

public_order_and_possession_of_weapons_rate_per_1_000_at_risk_population

Public order and Possession of weapons, rate per 1,000 'at-risk population'

anti_social_behaviour_rate_per_1_000_at_risk_population

Anti-Social Behaviour, rate per 1,000 'at-risk population'

connectivity_score

Connectivity Score

owner_occupation_affordability_component_of_housing_affordability_indicator

Owner-occupation affordability (component of housing affordability indicator)

private_rental_affordability_component_of_housing_affordability_indicator

Private rental affordability (component of housing affordability indicator)

housing_affordability_indicator

Housing affordability indicator

digital_connectivity_broadband_download_and_upload_speeds

Digital Connectivity (Broadband download and upload speeds)

core_homelessness_rate_percent_of_households

Core homelessness rate ( percent of households)

patient_to_gp_ratio

Patient-to-GP Ratio

homelessness_indicator_rate_per_1_000_households

Homelessness indicator (rate per 1,000 households)

household_overcrowding_indicator_rooms

Household overcrowding indicator (rooms)

household_overcrowding_indicator_bedrooms

Household overcrowding indicator (bedrooms)

...

Source

https://www.gov.uk/